[Auction Data] Convert unitPrice to BIGINT to support over 2 billion and add console.log of auctions length for row validation

+++ models/pricingdata.js
@@ -16,7 +16,7 @@ module.exports = (sequelize, DataTypes) => {
     }
   };
   pricingData.init({
-    unitPrice: DataTypes.INTEGER,
+    unitPrice: DataTypes.BIGINT,
     quantity: DataTypes.INTEGER
   }, {
     sequelize,
